bashrc: Remove upgrade file when system upgrade fails
It is important that the current date is written to the upgrade file
before the upgrade is run so that it won't be attempted to run when
spawning a new bash instance.
However I don't want the file to contain the current date if the
upgrade fails. So it has to be removed after the upgrade failed.